+++ include/templates.h
@@ -53,6 +53,7 @@ public:
     static void printVersion(Print &output);
     static void printWebInterfaceUrl(Print &output);
     static void printModel(Print &output);
+    static void printFileSystemInfo(Print &output);
 #if IOT_SSDP_SUPPORT
     static void printSSDPUUID(Print &output);
 #endif