+++ CHANGELOG.md
@@ -73,6 +73,7 @@
 - Customized WString class
 - Switch plugin fixes and improvements
 - Replaced the JSON library with a more stable version that consumes less memory
+- Switched to LittleFS
 
 ## Version 0.0.3